requirements.txt
nltk==3.1
oauthlib==1.0.3
PyGithub==1.26.0
requests==2.9.1
requests-oauthlib==0.6.0
twython==3.3.0
vaderSentiment==0.5
wheel==0.26.0
